+++ sybil/__init__.py
@@ -1,13 +1,12 @@
 from .sybil import Sybil
 from .document import Document
-from .region import Region, LexedRegion, Lexeme
+from .region import Region, Lexeme
 from .example import Example
 
 __all__ = [
     'Sybil',
     'Document',
     'Region',
-    'LexedRegion',
     'Lexeme',
     'Example',
 ]