+++ sybil/typing.py
@@ -1,4 +1,4 @@
-from typing import Callable, TYPE_CHECKING, Iterable, Optional, Any
+from typing import Callable, TYPE_CHECKING, Iterable, Optional, Any, Dict
 
 if TYPE_CHECKING:
     import sybil
@@ -8,7 +8,14 @@ if TYPE_CHECKING:
 Evaluator = Callable[['sybil.Example'], Optional[str]]
 
 #: The signature for a lexer. See :ref:`developing-parsers`.
-Lexer = Callable[['sybil.Document'], Iterable['sybil.LexedRegion']]
+#: Lexers must not set :attr:`~sybil.Region.parsed` or :attr:`~sybil.Region.evaluator`
+#: on the :class:`~sybil.Region` instances they return.
+Lexer = Callable[['sybil.Document'], Iterable['sybil.Region']]
 
 #: The signature for a parser. See :ref:`developing-parsers`.
 Parser = Callable[['sybil.Document'], Iterable['sybil.Region']]
+
+# In the future, this could likely be a TypedDict when Python 3.8 is the minimum supported version
+#: Mappings used to store lexemes for a :class:`~sybil.Region`.
+LexemeMapping = Dict[str, Any]
+